pyproject.toml
[project]
name = "spimex"
version = "0.1.0"
description = ""
authors = [
    {name = "Dmitry Longinov",email = "dl.ruled@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "alembic (>=1.16.4,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pandas (>=2.3.1,<3.0.0)",
    "xlrd (>=2.0.2,<3.0.0)",
    "beautifulsoup4 (>=4.13.4,<5.0.0)",
    "requests (>=2.32.4,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
